refactor(model) : 경로 계산 로직 추상화 해제
- Piece 에서 직접 구현
