+++ README.md
@@ -137,16 +137,16 @@
 
 - 궁성 구현 시 추가 고려 사항
     - 장
-        - [ ] 추가) 본인의 궁성 안을 나갈 수 없음 (팀에 따라 궁성 영역이 다름)
-        - [ ] 추가) 궁성 안의 간선(대각선)을 자유롭게 움직임.
+        - [x] 추가) 본인의 궁성 안을 나갈 수 없음 (팀에 따라 궁성 영역이 다름)
+        - [x] 추가) 궁성 안의 간선(대각선)을 자유롭게 움직임.
         - [x] 죽으면 게임이 끝남
     - 사
-        - [ ] 추가) 본인의 궁성 안을 나갈 수 없음 (팀에 따라 궁성 영역이 다름)
-        - [ ] 추가) 궁성 안의 간선(대각선)을 자유롭게 움직임.
+        - [x] 추가) 본인의 궁성 안을 나갈 수 없음 (팀에 따라 궁성 영역이 다름)
+        - [x] 추가) 궁성 안의 간선(대각선)을 자유롭게 움직임.
     - 차
         - [ ] 추가) 양 측 궁성 안의 간선(대각선)을 자유롭게 움직임.
     - 포
-        - [ ] 추가) 양 측 궁성 안의 간선(대각선)을![img.png](img.png) 자유롭게 움직임.
+        - [ ] 추가) 양 측 궁성 안의 간선(대각선)을 자유롭게 움직임.
     - 병
         - [ ] 상대 궁성에서 대각선 전진을 할 수 있음
