+++ src/main/java/model/piece/Ma.java
@@ -3,7 +3,6 @@ package model.piece;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.IntStream;
-import model.Path;
 import model.Point;
 import model.Team;
 
@@ -22,23 +21,6 @@ public class Ma extends Piece {
         return IntStream.range(0, horizontal.size())
                 .anyMatch(i -> horizontal.get(i) + beforePoint.x() == targetPoint.x()
                         && vertical.get(i) + beforePoint.y() == targetPoint.y());
-    }
-
-    @Override
-    public Path calculatePath(Point beforePoint, Point targetPoint) {
-        int vectorX = targetPoint.x() - beforePoint.x();
-        int vectorY = targetPoint.y() - beforePoint.y();
-
-        int unitVectorX = vectorX / Math.abs(vectorX);
-        int unitVectorY = vectorY / Math.abs(vectorY);
-
-        Point middlePoint = new Point(targetPoint.x() - unitVectorX, targetPoint.y() - unitVectorY);
-        Point endPoint = new Point(targetPoint.x(), targetPoint.y());
-
-        Path path = new Path();
-        path.addPoint(middlePoint);
-        path.addPoint(endPoint);
-        return path;
     }
 
     @Override
