+++ src/main/java/model/piece/goongsungpiece/Byeong.java
@@ -1,11 +1,12 @@
-package model.piece;
+package model.piece.goongsungpiece;
 
 import java.util.List;
 import java.util.Map;
 import java.util.stream.IntStream;
-import model.Path;
 import model.Point;
 import model.Team;
+import model.piece.Piece;
+import model.piece.PieceName;
 
 public class Byeong extends Piece {
 
@@ -28,13 +29,6 @@ public class Byeong extends Piece {
         return IntStream.range(0, horizontal.size())
                 .anyMatch(i -> horizontal.get(i) + beforePoint.x() == targetPoint.x()
                         && vertical.get(i) + beforePoint.y() == targetPoint.y());
-    }
-
-    @Override
-    public Path calculatePath(Point beforePoint, Point targetPoint) {
-        Path path = new Path();
-        path.addPoint(new Point(targetPoint.x(), targetPoint.y()));
-        return path;
     }
 
     @Override
