+++ src/main/java/model/piece/goongsungpiece/Jang.java
@@ -3,7 +3,7 @@ package model.piece.goongsungpiece;
 import model.Team;
 import model.piece.PieceName;
 
-public class Jang extends GoongsungPiece {
+public class Jang extends GoongsungConstrainedPiece {
 
     public Jang(Team team) {
         super(team);
