+++ src/main/java/model/piece/goongsungpiece/Pho.java
@@ -1,10 +1,11 @@
-package model.piece;
+package model.piece.goongsungpiece;
 
 import java.util.Map;
 import java.util.Map.Entry;
-import model.Path;
 import model.Point;
 import model.Team;
+import model.piece.Piece;
+import model.piece.PieceName;
 
 public class Pho extends Piece {
 
@@ -21,29 +22,6 @@ public class Pho extends Piece {
         return isStraightMove && !isSamePoint;
     }
 
-    @Override
-    public Path calculatePath(Point beforePoint, Point targetPoint) {
-        int vectorX = targetPoint.x() - beforePoint.x();
-        int vectorY = targetPoint.y() - beforePoint.y();
-
-        Path path = new Path();
-
-        if (vectorX == 0) {
-            int unitVectorY = vectorY / Math.abs(vectorY);
-            for (int i = 0; i < Math.abs(vectorY); i++) {
-                path.addPoint(new Point(targetPoint.x(), targetPoint.y() - unitVectorY * i));
-            }
-        }
-
-        if (vectorY == 0) {
-            int unitVectorX = vectorX / Math.abs(vectorX);
-            for (int i = 0; i < Math.abs(vectorX); i++) {
-                path.addPoint(new Point(targetPoint.x() - unitVectorX * i, targetPoint.y()));
-            }
-        }
-
-        return path;
-    }
 
     @Override
     public boolean canMove(Map<Piece, Boolean> piecesOnPathWithTargetOrNot) {
