+++ src/main/java/model/piece/goongsungpiece/Sa.java
@@ -3,7 +3,7 @@ package model.piece.goongsungpiece;
 import model.Team;
 import model.piece.PieceName;
 
-public class Sa extends GoongsungPiece {
+public class Sa extends GoongsungConstrainedPiece {
 
     public Sa(Team team) {
         super(team);
