+++ src/test/java/model/piece/ByeongTest.java
@@ -7,6 +7,8 @@ import java.util.HashMap;
 import java.util.Map;
 import model.Point;
 import model.Team;
+import model.piece.goongsungpiece.Byeong;
+import model.piece.goongsungpiece.Cha;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
