+++ src/test/java/model/piece/PhoTest.java
@@ -8,7 +8,9 @@ import java.util.Map;
 import model.Path;
 import model.Point;
 import model.Team;
+import model.piece.goongsungpiece.Cha;
 import model.piece.goongsungpiece.Jang;
+import model.piece.goongsungpiece.Pho;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
